requirements.txt
opencv-python
mediapipe
pyyaml